fix(generator): solve @commander-js/extra-typings not found error
The problem was caused by a missing dependency, @commander-js/extra-typings,
which was required by the neo4j-orm module's generator. This error prevented
the generator from working properly. I solved the issue by adding the missing
dependency to the package.json file and running 'npm install' to install it.

This commit fixes the error and restores the generator's functionality.

\# This commit addresses the following:
\# - Fixes the @commander-js/extra-typings not found error when using the generator CLI